add requiresRegionalTranslation
