+++ src/index.ts
@@ -4,7 +4,7 @@
 // ----- IMPORTS ----- //
 
 import _translateBatch from './api/batch/translateBatch';
-import _requiresTranslation, { _isSameDialect } from './locales/requiresTranslation';
+import _requiresTranslation from './locales/requiresTranslation';
 import _translate from './api/translate/translate';
 import _translateJsx from './api/jsx/translateJsx';
 import _updateProjectTranslations from './projects/updateProjectTranslations';
@@ -20,6 +20,7 @@ import { _isValidLocale, _standardizeLocale } from './locales/isValidLocale';
 import { _getLocaleName } from './locales/getLocaleName';
 import { _getLocaleDirection } from './locales/getLocaleDirection';
 import { defaultBaseUrl, libraryDefaultLocale } from './internal';
+import _isSameDialect from './locales/isSameDialect';
 
 // ----- HELPER FUNCTIONS ----- //
 
@@ -372,6 +373,28 @@ export function renderContentToString(content: Content, locales?: string | strin
 export function determineLocale(locales: string | string[], approvedLocales: string[]): string | undefined {
     return _determineLocale(locales, approvedLocales);
 };
+
+/**
+ * Determines whether a regional translation is required, (ie en-US to en-GB).
+ * 
+ * - If the target locale is not specified, the function returns `false`, as translation is not needed.
+ * - If the source and target language is the same, but the region is different return `true`.
+ * - If the `approvedLocales` array is provided, and the target locale is not within that array, the function also returns `false`.
+ * 
+ * @example `en-US` -> `en-GB` returns `true`
+ * @example `en-US` -> `en` returns `false`
+ * @example `en-GB` -> `en` returns `false`
+ * @example `en-US` -> `fr-FR` returns `false`
+ * 
+ * @param {string} sourceLocale - The locale code for the original content (BCP 47 locale code).
+ * @param {string} targetLocale - The locale code of the language to translate the content into (BCP 47 locale code).
+ * @param {string[]} [approvedLocale] - An optional array of approved target locales.
+ * 
+ * @returns {boolean} - Returns `true` if a regional translation is required, otherwise `false`.
+ */
+export function requiresRegionalTranslation(sourceLocale: string, targetLocale: string, approvedLocales?: string[]): boolean {
+    return _requiresTranslation(sourceLocale, targetLocale, approvedLocales);
+}
 
 /**
  * Determines whether a translation is required based on the source and target locales.
